performance: #682: Use Set, not List, when filtering possible genotypes
For particularly ambiguous searches, this has a fairly large effect!
The worst recorded case tested manually was reduced from 135 seconds to 68 seconds with a G-Group HF set, and from 728s to 469s in a g-group HF set. - so somewhere between a 35% and 50% improvement.
(Without the previously implemented genotype set truncation, match calculation dominated timings so much that these seconds would not be particularly significant)