chore: #682: Increase cap for expanded genotypes to 2000 for patient/donor
After experimentation with various cap sizes, the cap does not appear to have significant effect on the results of the algorithm above 1000. The higher the better in terms of accuracy, but returns are very diminishing.
As it stands, it seems that increasing from 1000-2000 has very little impact on performance, and 2000 therefore a better choice.
Doubling again to 4000 does detriment performance more noticeably, for non-noticeable accuracy gains, so no further increase is proposed at this time.